✨ Various upgrades and start of User's products
