+++ frontend/el-rastro/src/app/product/owner/[username]/layout.tsx
@@ -1,0 +1,13 @@
+export default function ProductLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  return (
+    <>
+      <main className="bg-gradient-to-tr from-white to-slate-200 min-h-screen">
+        {children}
+      </main>
+    </>
+  )
+}
