+++ frontend/el-rastro/src/app/product/owner/[username]/page.tsx
@@ -1,0 +1,8 @@
+import ProductList from "@/app/components/ProductsList"
+
+function ProductsOwner({ params }: { params: { username: string } }) {
+
+  return <ProductList activeOwner={params.username}/>
+}
+
+export default ProductsOwner
